Update types to be more compatible

+++ src/BKBase.h
@@ -53,8 +53,8 @@ typedef int32_t  BKInt;
 typedef uint32_t BKUInt;
 typedef int16_t  BKFrame;
 
-typedef ssize_t BKSize;
-typedef size_t  BKUSize;
+typedef long          BKSize;
+typedef unsigned long BKUSize;
 
 #define BK_FINT20_SHIFT 20
 #define BK_FINT20_UNIT (1 << BK_FINT20_SHIFT)
